requirements.txt
allure-python-commons~=2.13.2
selenium~=4.15.0
pytest~=7.4.3